+++ ext/tokenizers/src/pre_tokenizers.rs
@@ -1,26 +1,139 @@
+use std::sync::{Arc, RwLock};
+
+use serde::ser::SerializeStruct;
+use serde::{Deserialize, Serialize, Serializer};
+
 use tk::pre_tokenizers::bert::BertPreTokenizer;
+use tk::pre_tokenizers::whitespace::Whitespace;
+use tk::pre_tokenizers::PreTokenizerWrapper;
+use tk::{PreTokenizedString, PreTokenizer};
 
 #[magnus::wrap(class = "Tokenizers::PreTokenizer")]
-pub struct RbPreTokenizer {}
+#[derive(Clone, Serialize, Deserialize)]
+pub struct RbPreTokenizer {
+    #[serde(flatten)]
+    pub(crate) pretok: RbPreTokenizerTypeWrapper,
+}
+
+impl RbPreTokenizer {
+    #[allow(dead_code)]
+    pub(crate) fn new(pretok: RbPreTokenizerTypeWrapper) -> Self {
+        RbPreTokenizer { pretok }
+    }
+}
+
+impl PreTokenizer for RbPreTokenizer {
+    fn pre_tokenize(&self, normalized: &mut PreTokenizedString) -> tk::Result<()> {
+        self.pretok.pre_tokenize(normalized)
+    }
+}
 
 #[magnus::wrap(class = "Tokenizers::Whitespace")]
 pub struct RbWhitespace {}
 
 impl RbWhitespace {
-    pub fn new() -> Self {
-        Self {}
+    pub fn new() -> RbPreTokenizer {
+        Whitespace::default().into()
     }
 }
 
 #[magnus::wrap(class = "Tokenizers::BertPreTokenizer")]
-pub struct RbBertPreTokenizer {
-    pub pretok: BertPreTokenizer,
-}
+pub struct RbBertPreTokenizer {}
 
 impl RbBertPreTokenizer {
-    pub fn new() -> Self {
-        RbBertPreTokenizer {
-            pretok: BertPreTokenizer,
+    pub fn new() -> RbPreTokenizer {
+        BertPreTokenizer.into()
+    }
+}
+
+#[derive(Clone, Deserialize)]
+#[serde(untagged)]
+pub(crate) enum RbPreTokenizerWrapper {
+    // Custom(CustomPreTokenizer),
+    Wrapped(PreTokenizerWrapper),
+}
+
+impl Serialize for RbPreTokenizerWrapper {
+    fn serialize<S>(&self, serializer: S) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
+    where
+        S: Serializer,
+    {
+        match self {
+            RbPreTokenizerWrapper::Wrapped(inner) => inner.serialize(serializer),
+            // RbPreTokenizerWrapper::Custom(inner) => inner.serialize(serializer),
+        }
+    }
+}
+
+#[derive(Clone, Deserialize)]
+#[serde(untagged)]
+pub(crate) enum RbPreTokenizerTypeWrapper {
+    Sequence(Vec<Arc<RwLock<RbPreTokenizerWrapper>>>),
+    Single(Arc<RwLock<RbPreTokenizerWrapper>>),
+}
+
+impl Serialize for RbPreTokenizerTypeWrapper {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: Serializer,
+    {
+        match self {
+            RbPreTokenizerTypeWrapper::Sequence(seq) => {
+                let mut ser = serializer.serialize_struct("Sequence", 2)?;
+                ser.serialize_field("type", "Sequence")?;
+                ser.serialize_field("pretokenizers", seq)?;
+                ser.end()
+            }
+            RbPreTokenizerTypeWrapper::Single(inner) => inner.serialize(serializer),
+        }
+    }
+}
+
+impl<I> From<I> for RbPreTokenizerWrapper
+where
+    I: Into<PreTokenizerWrapper>,
+{
+    fn from(pretok: I) -> Self {
+        RbPreTokenizerWrapper::Wrapped(pretok.into())
+    }
+}
+
+impl<I> From<I> for RbPreTokenizerTypeWrapper
+where
+    I: Into<RbPreTokenizerWrapper>,
+{
+    fn from(pretok: I) -> Self {
+        RbPreTokenizerTypeWrapper::Single(Arc::new(RwLock::new(pretok.into())))
+    }
+}
+
+impl<I> From<I> for RbPreTokenizer
+where
+    I: Into<PreTokenizerWrapper>,
+{
+    fn from(pretok: I) -> Self {
+        RbPreTokenizer {
+            pretok: pretok.into().into(),
+        }
+    }
+}
+
+impl PreTokenizer for RbPreTokenizerTypeWrapper {
+    fn pre_tokenize(&self, pretok: &mut PreTokenizedString) -> tk::Result<()> {
+        match self {
+            RbPreTokenizerTypeWrapper::Single(inner) => inner.read().unwrap().pre_tokenize(pretok),
+            RbPreTokenizerTypeWrapper::Sequence(inner) => inner
+                .iter()
+                .try_for_each(|n| n.read().unwrap().pre_tokenize(pretok)),
+        }
+    }
+}
+
+impl PreTokenizer for RbPreTokenizerWrapper {
+    fn pre_tokenize(&self, pretok: &mut PreTokenizedString) -> tk::Result<()> {
+        match self {
+            RbPreTokenizerWrapper::Wrapped(inner) => inner.pre_tokenize(pretok),
+            // RbPreTokenizerWrapper::Custom(inner) => inner.pre_tokenize(pretok),
         }
     }
 }