+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.3.0 (unreleased)
 
+- Added support for training tokenizers
 - Added `pair` argument to `encode` method
 - Changed `encode` method to include special tokens by default
 - Changed how offsets are calculated for strings with multibyte characters